TestResult: added
Added Test Result struct containing test duration and  a std::vector with the intersection results. This is returned from the run test function in Intersection test.

+++ DV2556/DV2556/DV2556/StopWatch.cpp
@@ -8,7 +8,7 @@ void StopWatch::start()
 float StopWatch::getTimeInSeconds()
 {
 	clock_t difference = clock() - s;
-	return difference / CLOCKS_PER_SEC;
+	return (float)difference / CLOCKS_PER_SEC;
 }
 
 int StopWatch::getTimeInMilliSeconds()
